app/requirements.txt
streamlit
geopandas
folium
streamlit-folium
shapely
pyproj
pandas
mysql-connector-python